backend/pyproject.toml
[tool.poetry]
name = "LuckySite_test_project"
version = "0.1.0"
description = ""
authors = ["borsh130894 borsh130894@gmail.com"]
packages = [
    {include = "main"}
]

[tool.poetry.dependencies]
python = ">=3.10,<4"
django = ">=5.2.2,<6.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
django-environ = "^0.12.0"
pillow = "^11.2.1"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"